Reworked Game Size and Random Game
Changed the game resize option to be done on startup so the program loads a little longer but works faster in general.
Also changed the random game list functionality to be a single random game at a time with the UI as well.
Last small change: managed the single game view functionality making it so that a single game is visible from the start on some selections.